# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Healthy Nation API — vitals, doctors, medications, AI triage, emergency, pharmacy
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vitals:
    get:
      operationId: listVitals
      tags: [vitals]
      summary: Recent vital readings
      parameters:
        - in: query
          name: vitalType
          schema:
            $ref: "#/components/schemas/VitalType"
        - in: query
          name: days
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Readings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VitalReading"
    post:
      operationId: createVital
      tags: [vitals]
      summary: Record a vital reading
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVitalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VitalReading"
...
  /vitals/trends:
    get:
      operationId: getVitalTrends
      tags: [vitals]
      summary: Time-series points for chart rendering
      parameters:
        - in: query
          name: vitalType
          required: true
          schema:
            $ref: "#/components/schemas/VitalType"
        - in: query
          name: days
          schema:
            type: integer
            default: 14
      responses:
        "200":
          description: Trend points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VitalTrendPoint"
...
  /medicines:
    get:
      operationId: listMedicines
      tags: [pharmacy]
      summary: Browse / search the medicine catalog
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
      responses:
        "200":
          description: Medicines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medicine"
  /medicines/{id}:
    get:
      operationId: getMedicine
      tags: [pharmacy]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Medicine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
...
  /orders:
    get:
      operationId: listOrders
      tags: [pharmacy]
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [pharmacy]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [pharmacy]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
...
components:
  schemas:
...
    VitalType:
      type: string
      enum: [heart_rate, spo2, blood_pressure, glucose, temperature, steps, sleep, respiratory_rate]
...
    OrderStatus:
      type: string
      enum: [pending_payment, preparing, dispatched, out_for_delivery, delivered, cancelled]
...
    VitalReading:
      type: object
      required: [id, vitalType, recordedAt, value, unit, status]
      properties:
        id:
          type: integer
        vitalType:
          $ref: "#/components/schemas/VitalType"
        recordedAt:
          type: string
          format: date-time
        value:
          type: number
        valueSecondary:
          type: ["number", "null"]
        unit:
          type: string
        status:
          type: string
          enum: [normal, warning, critical]
        deviceName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    CreateVitalInput:
      type: object
      required: [vitalType, value]
      properties:
        vitalType:
          $ref: "#/components/schemas/VitalType"
        value:
          type: number
        valueSecondary:
          type: ["number", "null"]
        unit:
          type: ["string", "null"]
        deviceName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...
    VitalTrendPoint:
      type: object
      required: [timestamp, value]
      properties:
        timestamp:
          type: string
          format: date-time
        value:
          type: number
        valueSecondary:
          type: ["number", "null"]
...
    Medicine:
      type: object
      required: [id, name, strength, form, manufacturer, price, mrp, inStock, requiresPrescription, category, description]
      properties:
        id:
          type: integer
        name:
          type: string
        strength:
          type: string
        form:
          type: string
        manufacturer:
          type: string
        price:
          type: integer
        mrp:
          type: integer
        inStock:
          type: boolean
        requiresPrescription:
          type: boolean
        category:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
...
    OrderItemInput:
      type: object
      required: [medicineId, quantity]
      properties:
        medicineId:
          type: integer
        quantity:
          type: integer
    CreateOrderInput:
      type: object
      required: [pharmacyId, items, deliveryAddress, paymentMethod]
      properties:
        pharmacyId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
        deliveryAddress:
          type: string
        paymentMethod:
          type: string
          enum: [card, upi, cod, wallet]
        notes:
          type: ["string", "null"]
    OrderItem:
      type: object
      required: [medicineId, name, strength, quantity, unitPrice]
      properties:
        medicineId:
          type: integer
        name:
          type: string
        strength:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: integer
    Order:
      type: object
      required: [id, status, pharmacyName, totalAmount, itemCount, createdAt, estimatedDeliveryAt, currentStep]
      properties:
        id:
          type: integer
        status:
          $ref: "#/components/schemas/OrderStatus"
        pharmacyName:
          type: string
        totalAmount:
          type: integer
        itemCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        estimatedDeliveryAt:
          type: string
          format: date-time
        currentStep:
          type: integer
    OrderDetail:
      type: object
      required: [id, status, pharmacyName, items, subtotal, deliveryFee, tax, totalAmount, createdAt, estimatedDeliveryAt, deliveryAddress, paymentMethod, currentStep, steps]
      properties:
        id:
          type: integer
        status:
          $ref: "#/components/schemas/OrderStatus"
        pharmacyName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: integer
        deliveryFee:
          type: integer
        tax:
          type: integer
        totalAmount:
          type: integer
        createdAt:
          type: string
          format: date-time
        estimatedDeliveryAt:
          type: string
          format: date-time
        deliveryAddress:
          type: string
        paymentMethod:
          type: string
        currentStep:
          type: integer
        driverName:
          type: ["string", "null"]
        driverPhone:
          type: ["string", "null"]
        steps:
          type: array
          items:
            $ref: "#/components/schemas/OrderStep"
    OrderStep:
      type: object
      required: [label, status]
      properties:
        label:
          type: string
        status:
          type: string
          enum: [done, current, pending]
        timestamp:
          type: ["string", "null"]
...