# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Healthy Nation API — vitals, doctors, medications, AI triage, emergency, pharmacy
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  status: OrderStatus;
  pharmacyName: string;
  totalAmount: number;
  itemCount: number;
  createdAt: Date;
  estimatedDeliveryAt: Date;
  currentStep: number;
}