# lib/api-zod/src/generated/types/vitalTrendPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Healthy Nation API — vitals, doctors, medications, AI triage, emergency, pharmacy
 * OpenAPI spec version: 0.1.0
 */

export interface VitalTrendPoint {
  timestamp: Date;
  value: number;
  /** @nullable */
  valueSecondary?: number | null;
}